# documentation/static/openapi.yml.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:5000
    description: Generated server url
...
paths:
  /webaccount/socialintegration:
    get:
      tags:
        - Web Account Service
      summary: Registers WebAccounts during PatreonShout sign up
      operationId: GetSocialIntegration
      parameters:
        - name: login_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Integrations returned
          content:
            application/json: {}
    put:
      tags:
        - Web Account Service
      summary: Registers WebAccounts during PatreonShout sign up
      operationId: PutSocialIntegration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutSocialIntegrationRequest'
        required: true
      responses:
        '200':
          description: Integration has been saved
          content:
            application/json: {}
        '409':
          description: WebAccount ID does not exist
          content:
            application/json: {}
...
  /webaccount/webhook:
    post:
      tags:
        - Web Account Service
      operationId: Webhook
      parameters:
        - name: login_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /receivers/patreon/webhook/{webaccountId}:
    post:
      tags:
        - Receiver Service
      summary: >-
        Receives post information from Patreon after a post is either published,
        updated, or deleted
      operationId: PatreonWebhook
      parameters:
        - name: x-patreon-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-patreon-event
          in: header
          required: true
          schema:
            type: string
        - name: User-Agent
          in: header
          required: true
          schema:
            type: string
        - name: webaccountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
        required: true
      responses:
        '200':
          description: none (patreon doesn't care)
          content:
            '*/*':
              schema:
                type: object
...
  /receivers/twitter/oauth:
    get:
      tags:
        - Receiver Service
      summary: Receives information from Twitter after OAuth
      operationId: TwitterOAuth
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: HTML message saying 'Twitter linked!'
          content:
            '*/*':
              schema:
                type: string
  /receivers/patreon/oauth:
    get:
      tags:
        - Receiver Service
      summary: Receives information from Patreon after OAuth
      operationId: PatreonOAuth
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: HTML message saying 'Patreon linked!'
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    PutSocialIntegrationRequest:
      type: object
      properties:
        login_token:
          type: string
        integration_name:
          type: string
          enum:
            - DISCORD
            - TWITTER
            - INSTAGRAM
        data:
          type: string
...
    PatreonCampaignV2:
      type: object
      properties:
        vanity:
          type: string
        data:
          $ref: '#/components/schemas/PatreonDataV2'
    PatreonDataV2:
      type: object
      properties:
        attributes:
          type: object
        id:
          type: integer
          format: int32
        relationships:
          $ref: '#/components/schemas/PatreonRelationshipsV2'
        type:
          type: string
    PatreonLinksV2:
      type: object
      properties:
        self:
          type: string
    PatreonRelationshipsV2:
      type: object
      properties:
        campaign:
          $ref: '#/components/schemas/PatreonCampaignV2'
    WebhookRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PatreonDataV2'
        links:
          $ref: '#/components/schemas/PatreonLinksV2'
...